* 支撑达梦数据库 * 支撑内置sqlite数据库，做到无数据库使用 * 国标设备推拉流 统一一张表 * 增加AK/SK 认证，给业务系统调用接口

+++ ruoyi-wvp/src/main/java/com/ruoyi/wvp/controller/CloudRecordController.java
@@ -428,7 +428,7 @@ public class CloudRecordController extends BaseController {
         if (!ObjectUtils.isEmpty(list)) {
             for (CloudRecordItem cloudRecordItem : list) {
                 CloudRecordUrl cloudRecordUrl = new CloudRecordUrl();
-                cloudRecordUrl.setId(cloudRecordItem.getId());
+                cloudRecordUrl.setId(cloudRecordItem.getId().intValue());
                 cloudRecordUrl.setDownloadUrl(remoteHost + "/index/api/downloadFile?file_path=" + cloudRecordItem.getFilePath() + "&save_name=" + cloudRecordItem.getStream() + "_" + cloudRecordItem.getCallId() + "_" + DateUtil.timestampMsToUrlToyyyy_MM_dd_HH_mm_ss(cloudRecordItem.getStartTime()));
                 cloudRecordUrl.setPlayUrl(remoteHost + "/index/api/downloadFile?file_path=" + cloudRecordItem.getFilePath());
                 cloudRecordUrlList.add(cloudRecordUrl);
